requirements.txt
box2d-py==2.3.8
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
Click==7.0
Flask==1.0.2
future==0.17.1
gym==0.12.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mkl-fft==1.0.10
mkl-random==1.0.2
numpy==1.16.2
Pillow==6.0.0
protobuf==3.7.1
pycparser==2.19
pyglet==1.3.2
requests==2.21.0
scipy==1.2.1
six==1.12.0
torch==1.0.1.post2
urllib3==1.24.1
visualdl==1.3.0
Werkzeug==0.15.2
